[REFACTOR(BridgeRandomNumberGenerator,BridgeNumberGenerator)] 패키지구조 변경
