+++ src/main/kotlin/bridge/domain/BridgeNumberGenerator.kt
@@ -1,4 +1,4 @@
-package bridge
+package bridge.domain
 
 interface BridgeNumberGenerator {
     fun generate(): Int
